Language.Janus.Interp

Documentation

data EvalState

Intepreter state.

data EvalError

Interpreter error.

Constructors

OpCallTypeError String [[TypeRep]] [TypeRep]
Operator invocation type error

ItemCallError Item EvalError
Callable (e.g. function) call error

IndexOutOfBounds
Tried to access out of bounds index

InternalError String
Internal interpreter error - a bug

InvalidPointer Ptr
Tried to access unallocated memory cell

ExpectedBool Val
Expected boolean value (e.g. in condition)

ExpectedRef
Expected reference (e.g. in call operator)

NotCallable Val
Given value is not callable

OutOfMemory
Cannot allocate more memory cells

UndefinedSymbol String
Cannot resolve requested symbol

CustomError String
In-code exception

type InterpM = StateT EvalState (ExceptT EvalError IO)

Interpreter monad.

runInterpM :: InterpM a -> IO (Either EvalError a)

Run action inside clean InterpM.

run :: Evaluable a => a -> IO (Either EvalError Val)

Evaluate Evaluable entity inside clean InterpM.

deref :: Ref -> InterpM Val

Dereference a Ref - get Val given reference points to.

refset :: Ref -> Val -> InterpM ()

Alter value given Ref points to.

pushScope :: InterpM ()

Push new scope.

popFrame :: InterpM ()

Pop top stack frame.

memIsFree :: Ptr -> InterpM Bool

Check whether given pointer points to unallocated memory cell.

memGetVal :: Ptr -> InterpM Val

Get value of given memory cell.

memGetRc :: Ptr -> InterpM Word

Get reference count of given memory cell.

malloc :: Val -> InterpM Ptr

Allocate value in interpreter memory. Returns new memory cell's address.

memset :: Ptr -> Val -> InterpM ()

Set value at given address.

rcIncr :: Ptr -> InterpM ()

Increment reference count of memory cell at given address.

rcDecr :: Ptr -> InterpM ()

Decrement reference count of memory cell at given address. If the RC will reach 0, the cell will be deallocated.

lookupVar :: String -> InterpM Ptr

Lookup variable in all available scopes.

putVar :: String -> Ptr -> InterpM ()

Create new, or change existing, variable binding. If the variable has been already exising, its old value's reference count will be decremented.

evalVal :: String -> InterpM Val

   evalVal name = lookupVar name >>= memGetVal

allVars :: InterpM [String]

Enumerate all visible symbols.

class Evaluable a where

The Evaluable class denotes enitity which can be evaluated to Val.

Minimal complete definition

eval

Methods

eval :: a -> InterpM Val

class RefEvaluable a where

The RefEvaluable class denotes enitity which can be evaluated to Ref

Minimal complete definition

evalRef | tryEvalRef

Methods

evalRef :: a -> InterpM Ref

Throws ExpectedRef if tryEvalRef returns Nothing.

tryEvalRef :: a -> InterpM (Maybe Ref)

class Callable a where

The Callable class denotes enitity which can be called, such as functions.

Minimal complete definition

call

Methods

call :: a -> [Val] -> InterpM Val

valGetIdx

Arguments

:: Val
Container value

-> Val
Index

-> InterpM Val
Value inside container at given index

valSetIdx

Arguments

:: Val
Container value

-> Val
Index

-> Val
New Value

-> InterpM Val
New container with value at given index changed